[refactor][storage][badger]Refactored the prefilling of cache to reader
